Offer a default StripSecrets for playerView
Adds a game Liar's Dice to the example demos, which shows that PlayerView
will strip the secrets from a game's state for a player.

Also, makes a change to the local Client, which automatically applies
PlayerView, even if the default `(G) => G` reducer is used. I guess
this is a bug, but it's probably most of the time you don't want to do
client-side secret stripping, and most of the time this is for server.