Add redux enhancer option
Adds an optional param to a Client which allows the user to hook into the Redux store with an enhancer or middleware.

The main motivation for this is to hook into it with [`redux-logger`](https://github.com/evgenyrodionov/redux-logger) or [Redux Devtools](http://extension.remotedev.io/), and is documented in an additional page.