separate log object and socket events that modify it
Split the socket events into:

1. SYNC

This is called at the beginning of the connection. The server
sends back the entire state and the entire log.

2. UPDATE

This is called after each move (it is broadacast to all clients).
The server sends back a deltalog and the entire state.

The log object is maintained in client.js via Redux middleware.

On the server, it is still kept inside the state object, although
a future change will split it into a separate area of storage
as an optimization.