change phases syntax
This is in preparation for an upcoming feature to quickly switch
to a temporary phase (and back after it is done).

OLD:
phases: [
  { name: 'A', ... }
  { name: 'B', ... }
]

NEW:
phases: {
  A: { ... },
  B: { ... },
}

This allows creation of different phase orderings that aren't
simple lists. Each phase can now point to a `next` phase:

phases: {
  A: { next: 'B' },
}

A 'default' phase is the catch-all next phase for phases that don't
specify `next`.

In order to emulate the old behavior, you can do something like:

phases: {
  A: { next: 'B' },
  B: { next: 'C' },
  C: { next: 'A' },
}