Add adminClient option for Firebase storage
I expect this to be backward-compatible with all existing users.  The
only gotcha might be people blacklisting dynamic "require"s for webpack;
they'll need to also blacklist 'firebase-admin'.

Add documentation for the new and existing properties to the tutorial.
Also updates the example to use Firestore and adminClient.

Fixes nicolodavis/boardgame.io#384