test: `turn.stages` not applied to `ctx.stage` on new turn
When initialising a phase, `boardgame.io` uses the `turn.stages` object 
to set up `ctx.stage` for all players. However, given that `stages` is a 
property of `turn`, one might expect `stages` to be used to update 
`ctx.stage` also when a turn starts/ends, e.g. to update `currentPlayer` 
to the correct stage.