refactor(client): Scope global CSS selectors in Debug panel
Previously, `<button>`, `<section>`, and `<select>` elements where being 
styled directly with global scope. This poses a fairly high risk of 
leaking styles to elements in the rest of a game using the debug panel.

This restricts the scope of those global selectors to only apply to 
children of the root debug panel component.