fix(master): Remove credentials from action payloads after use
Credentials are only used to check if an action is authorized in the 
master and should not then be passed on to the reducer etc.

This contributes towards #227, because now credentials should not leak 
beyond the `Master.onUpdate` method and won’t end up for example in the 
game log.