feat(plugins): Make player plugin a factory function
This moves the `playerSetup` function out of the game configuration 
object. Instead `PluginPlayer` should be called with the setup function 
as its argument and it will return the plugin object boardgame.io 
expects.

BREAKING CHANGE: The `PluginPlayer` is no longer a plain object that 
looks for game.playerSetup during its setup. Instead, you have to pass 
your `playerSetup` method directly to the plugin.

Closes #603.