Include credentials in undo/redo actions
If playing on a remote server, actions are validated wrt player
credentials that are assumed to be inside action.payload.credentials:
https://github.com/nicolodavis/boardgame.io/blob/v0.39.3/src/master/master.ts#L188

However, in case of undo and redo, the client sends an action that
consists only of type:
https://github.com/nicolodavis/boardgame.io/blob/v0.39.3/src/core/action-creators.ts#L97

Make it possible to undo/redo with a remote server and fix #593 by
adding credentials to undo/redo actions.

All the remaining actions without credentials are client-only, so remove
them from CredentialedActionShape and make the master only accept
CredentialedActionShape.Any in onUpdate().