build: Fixes installing boardgame.io as a git dependency
When installing as a git dependency, e.g. `npm i 
nicolodavis/boardgame.io`, the prepack script is run but will error 
because development dependencies aren’t available. This adds an `npm 
install` step to the prepack pipeline, which admittedly will generally 
be redundant, but I don’t see a way to only run it for git dependency 
installs and it shouldn’t cause any unwanted behaviour in other 
scenarios.