fix(types): Import Svelte types in client instead of top-level types
We import Svelte to hoist its ambient '*.svelte' module type 
declaration. However, this depends on browser/DOM types, which TS 
consumers not implementing for the browser may not have active (e.g. if 
building a storage connector or plugin), so it makes sense to only 
import this in the client itself where consumers will need `"lib": 
["dom"]` in their tsconfig.json or equivalent.