test(debug): Mock `svelte-json-tree-auto` in tests
`svelte-json-tree-auto` causes errors in tests (see #781). I think this 
might have something to do with the fact that the tree has circular 
dependencies, which perhaps the Jest transformer can’t handle. As it’s a 
third-party library, we can probably do without it during tests, and 
this change replaces it with an empty Svelte component for tests.