fix(debug): Use latest moves & events dispatchers
Closes #771

Previously, the debug panel only read the moves & events dispatchers 
from the client when instantiated. This prevented users from toggling 
between players and dispatching moves from the debug panel as different 
players with stages, because the move/event would be out-of-date 
(created with the wrong playerID). This change makes sure we’re always 
referencing the latest dispatchers.