feat(debug): Add ClientManager to track client & debug panel instances
Fixes #810, closes #670

This adds a global client manager which all clients register and 
unregister with when they start/stop. Mounting/unmounting the debug 
panel is now delegated to this manager, which allows us to only ever 
render one debug panel. To handle this, the debug panel now displays a 
client switcher when multiple clients are in use. Toggling the players 
will also try to switch to a client for that player before it actually 
updates the current client.