refactor(master): Don’t use sync storage == sync auth heuristic
Previously `authenticateCredentials` was awaited if the master was using 
an async storage implementation, but not when the master was using 
synchronous storage. This was so as to avoid `await` when running the 
master locally where it can be treated as a synchronous API. This 
assumption could be buggy though if a user provides an async 
`authenticateCredentials` method (e.g. to call out to a third-party auth 
service), but is using the default synchronous `InMemory` storage.

This commit refactors the authentication logic to assume that if the 
master has been provided with an `Auth` instance, it can be run 
asynchronously. Local masters will run without an `Auth` instance, so 
will skip authentication entirely.