ci(Github): Only run publish workflow on tags
I set up the publish workflow with both `branches` and `tags` options, thinking that would set tags *on* the specified branches. Instead it runs the workflow for both the specified branches AND the specified tags. This commit removes the `branches` option to only run the publish workflow for pushed tags.