refactor: Enable TransportAPI to send to a provisional client
A provisional client is a client that hasn’t yet been authenticated by 
the game master. Adding it to `clientInfo` could be risky becase it 
could receive payloads intended for a specific playerID before it has 
been authenticated for that playerID. With this approach, the 
provisional client is sandboxed until it is authenticated.