fix: Improve security of socket.io transport layer (#894)
* test: Restructure simultaneous socket test

* test: Add ability to override random delays in async test storage

* test: Write tests cases for inauthentic client connection

This implements several tests for the data leaking scenarios discussed 
in #889

* refactor: Remove unnecessary use of `apply`

* refactor: Add addClient/removeClient methods for clarity and consistency

* fix: Remove clients if the master reports they are inauthentic

This addresses inauthentic clients receiving a sync triggered by a later 
authentic client’s sync (as described in #889).

* refactor: Extract the emit logic in TransportAPI

* refactor: Don’t use `send` in socket transport’s `sendAll`

Both `send` and `sendAll` use loops to iterate over the registered 
clients and emit events. Using `sendAll` would emit an exponential 
number of events if multiple clients were connected with the same 
`playerID`.

* refactor: Make `send` and `sendAll` consistent with each other

* style: Add blank line between event handlers

* refactor: Enable TransportAPI to send to a provisional client

A provisional client is a client that hasn’t yet been authenticated by 
the game master. Adding it to `clientInfo` could be risky becase it 
could receive payloads intended for a specific playerID before it has 
been authenticated for that playerID. With this approach, the 
provisional client is sandboxed until it is authenticated.

* fix: Only add clients to clientInfo after initial authentication

This prevents a race condition whereby if an inauthentic client synced 
just before an authentic client, but its authentication process took 
longer than the authentic client’s sync, it could receive the authentic 
client’s sync payload.

* refactor: Remove unnecessary `removeClient` call

* refactor: Destructure variables in addClient for readability

* fix: Clean up `roomInfo` entries for empty matches

Memory overhead for a bunch of empty Sets is probably marginal, but it 
seems like these should be removed nonetheless.

* refactor: Rename variable

* style: Whitespace and variable rename