fix(lobby): Clean up/update lobby polling interval properly
Previously, the lobby component started polling on an interval when created, but that interval was never cleared, resulting in the polling continuing even after the component unmounted.

Similarly, the lobby component receives a `refreshInterval` prop, but we weren’t ever updating the polling interval if that prop changed. This commit also updates the polling interval correctly.